Move extension to own file
